Rename runtimeOps Tree
rename runtime ops to environment

contributes to #1088
